requirements.txt
altgraph==0.17.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.5
astroid==2.3.3
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
backports.zoneinfo==0.2.1
base==1.0.4
beautifulsoup4==4.8.2
bleach==3.1.0
blinker==1.4
boto3==1.14.63
botocore==1.17.63
cached-property==1.5.1
cachetools==4.2.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.1.1
click==7.0
clickhouse-driver==0.1.4
cloudflare==2.8.15
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
datefinder==0.7.3
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.6
distro==1.4.0
setuptools==65.3.0
Django==3.0.14
docker==4.1.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.4
Flask-Cors==3.0.10
websocket-client==0.53.0
urllib3==1.24.3
selenium==3.141.0
Flask-Excel==0.0.7
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-Restless==1.0.0b1
future==0.18.2
futures==2.2.0
geocoder==1.38.1
geoip2==2.9.0
googlesearch-python==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image==1.5.28
importlib-metadata==4.10.0 
importlib-resources==5.10.0
incremental==16.10.1
infi.clickhouse-orm==2.1.0
inflect==4.0.0
ip2geotools==0.1.5
IP2Location==8.4.1
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jeepney==0.8.0
Jinja2==2.10.3
jira==3.0.1
jmespath==0.10.0
jsonlines==1.2.0
jsonpatch==1.23
jsonpickle==2.0.0
jsonpointer==2.0
jsonurl==1.0.0
keyring==19.3.0
kiwisolver==1.1.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
maxminddb==1.5.4
maxminddb-geolite2==2018.703
mccabe==0.6.1
mimerender==0.6.0
more-itertools==8.0.2
msc-pyparser==1.2.1
requests==2.28.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
aniso8601==8.0.0
mysql-connector-python==8.0.18
mysqlclient==1.4.6
Naked==0.1.31
netaddr==0.7.19
netifaces==0.10.9
netplan==0.3.3
nginxparser==1.0.1
ntplib==0.3.4
numpy==1.18.5
oauthlib==3.1.1
packaging==21.3
pexpect==4.6.0
pika==1.3.1
Pillow==7.0.0
pip-review==1.0
pkginfo==1.9.6
pkgutil-resolve-name==1.3.10
ply==3.11
protobuf==3.10.0
psutil==5.7.3
publicsuffixlist==0.9.3
py-cpuinfo==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.19
pycryptodome==3.9.9
pycti==5.3.17
pycurl==7.44.1
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.5.2
PyHamcrest==1.9.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
pymisp==2.4.162.2
PyNaCl==1.3.0
pyotp==2.6.0
pyparsing==2.4.6
pyquery==1.4.1
pyrsistent==0.15.7
pyserial==3.4
pyTelegramBotAPI==4.1.1
python-dateutil==2.8.2
python-debian==0.1.36
python-dotenv==0.10.5
python-iptables==0.14.0
python-json-logger==2.0.4
python-magic==0.4.27
python-mimeparse==1.6.0
python-nginx==1.5.3
python-telegram-bot==13.15
pythonping==1.0.8
pytz==2019.3
PyYAML==6.0
ratelim==0.1.6
readme-renderer==24.0
redis==4.3.5
regex==2022.9.13
s3transfer==0.3.3
schedule==0.6.0
SecretStorage==3.3.1
service-identity==18.1.0
shellescape==3.4.1
simplejson==3.16.0
six==1.13.0
sos==0.23.4
soupsieve==2.0.1
SQLAlchemy==1.3.24
sqlparse==0.3.1
sseclient==0.0.27
ssh-import-id==5.11
stix2==3.0.1
stix2-patterns==2.0.0
systemd-python==234
tabulate==0.8.10
taxii2-client==2.3.0
texttable==1.6.2
tornado==6.1
tqdm==4.40.2
twilio==6.40.0
twine==3.1.1
Twisted==18.9.0
typed-ast==1.4.3
typing==3.7.4.3
tzlocal==2.1
validators==0.14.2
virtualenv==20.0.35
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.11.2
zipp==3.9.1
zope.interface==4.7.2
Werkzeug==0.16.1
jsonschema==4.17.3